feat(ui): update tapplet action (#111)
Description
---
Adds update functionality to installed tapplets.

Motivation and Context
---
Basic functionality for the users.

How Has This Been Tested?
---
Tested manually

What process can a PR reviewer use to test or verify this change?
---
Use tapp-registrant to update version and click a button to verify.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->